requirements.txt
Brotli @ file:///D:/bld/brotli-split_1687884251065/work
certifi==2023.7.22
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1688813409104/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1691577114857/work
numpy @ file:///D:/bld/numpy_1691056618426/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1681337016113/work
pandas @ file:///D:/bld/pandas_1688740748520/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1690813113769/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1679580333621/work
PySocks @ file:///D:/bld/pysocks_1661604991356/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1626286286081/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1680088766131/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1684774241324/work
scikit-learn @ file:///D:/bld/scikit-learn_1688116465061/work
scipy==1.11.2
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
sklearn-relief==1.0.0b2
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1689261241048/work
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1688315532570/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1680081134351/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1689789803562/work
win-inet-pton @ file:///D:/bld/win_inet_pton_1667051142467/work
